GEODE-9730: release script improvements (#6989)
* add fan-in for rc pipeline
* disable parallel when publishing to nexus to avoid possibility of upload getting split into two repos
* switch native build from debian to ubuntu for a (less) brittle was to get a recent-enough version of cmake (min 3.12, currently this gets 3.16)
* fix to support pinned geode sha, fix for concourse getting confused and seeing the wrong tag
* Don't imply there's anything to backport for other native branches unless this is a flagship release
* provide instructions to check for and recover from split repo
* remind that issues fixed in e.g. 1.13.1 and 1.14.1 should not be closed if 1.13.1 is released first; wait to close until latest release has fix
* maintain geode-benchmarks versionNumber, even though it doesn't seem to be used for anything
* maintain default benchmark baseline on a new flagship release
* tighten expected file sizes
* template the contents of the GitHub release description
* gpg has been split into two packages
* include a reminder to write the releasenotes while the rc pipeline is running before sending the vote email
* increase upthewaterspout timeout to 2h after observing it can take longer than 1h
* fix benchmark baseline updating and update to the minor for the support branch, unlike develop
* improve prompts for creating release on GitHub

+++ dev-tools/release/license_review.sh
@@ -132,6 +132,7 @@ function listJarsInWar() {
 }
 
 function extractLicense() {
+  [ "$SKIP_LICENSES" != "true" ] || return 0
   war=$1
   rm -Rf tmpl
   mkdir tmpl
